extract Search, Category Handler
